ugly hacks to workaround which seems to be nasty Android bugs
When encoding to AAC, Android AAC MediaCodec inserts wrong blank frames at the beginning of the sound and trims the end.
The workaround: don't write the first 2 buffers from the encoder, and append 2048 0-samples at the end of the PCM stream which feeds the encoder.

Alternatively I also could be a stupid coder, not sufficiently helped by what exists as documentation, and I do it all wrong.